Implements both part NN and Main Menu/Back in detection.
  The part NN is fairly simple due to the 'Continued' event written as
the last line of the previous file.

  The Main Menu detection isn far from perfect due to there being no
clear event written when the player exits to there.  There *is* a
"Music" event, but it's with "MusicTrack":"Exploration", not "MainMenu".
The "Exploration" music track is also played, at least, when docked at
(some?) stations.
  So we're fudging it by looking for a "Commander" event after the
"Location" event we use to find the start of a session, and then using
the timestamp of the event/line immediately before that.
  This works because we're using that "Location" event as the start of a
session, and the "Commander" event for that session is written *before*
the "Location" event, as the first logging in event.

  NB: This still won't handle:

  	1. Player logs in, plays for a while or not.
	2. Player exits to main menu and idles there.
	3. Player fully exits the game without logging in to the game
	   proper again.